+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.11.0
+
+- `association` and `associationOrFail` handle required and passthrough `and` clauses
+- `association` and `associationOrFail` compatible with transactions
+
 ## 1.10.0
 
 - bump kysely